feat: Add version dropdown for selecting different versions
The code changes introduce a new `VersionDropdown` component that allows users to select different versions of the application. This component fetches the available versions from the `versions.json` file and renders them as options in a dropdown menu. When a version is selected, the page is redirected to the corresponding URL.